requirements.txt
flask
numpy
faiss-cpu
sentence-transformers
